Cell Phone Part 1
Player info works, GPS is there but need to implement switching between the two.

+++ Assets/Scripts/Player Remake/HealthAndRespawn.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class HealthAndRespawn : MonoBehaviour
 {
-    public float health = 100;
+    public int health = 100;
 
     public GameObject checkpoint;
